overlay-octopus/chromeos-bsp-octopus: Enable stylus tools
This change is used to enable stylus tools while support pen eject as
wake source. This change also enables rules for powerd to decide when to
enable pen eject as wakeup source.

BRANCH=octopus
BUG=b:117953118
TEST=Ensure that the system boots to ChromeOS. Ensure that the stylus
tools open on pen eject. Ensure that the system can enter S0ix and S3
states successfully when the pen is inserted. Ensure that the system
wakes on Pen Eject. Ensure that the system enters S0ix and S3 states
when the pen is placed in its holder. Ensure that the suspend_stress_test
runs successfully for 25 iterations with the pen placed in its holder
and pen ejected from its holder.

Change-Id: I003acc50e819d18b31cff27b0787a95e0385896d
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1368568
Commit-Ready: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>