octopus: Use a generic filename for cras settings
Enabling SOF adds an "sof-" prefix to the audio card name. This causes
the volume settings in cras-config to be lost when enabling SOF.

Instead of renaming the underlying files, this change uses a generic
filename inside of cras-config for storing these card settings. I had
initially renamed them, but that would make it hard to do any
experimentation with SOF disabled, if need be.

Talking to cujomalainey@, it sounds like these parameters should be the
same whether or not SOF is enabled, so we should be okay using a generic
file here.

BUG=b:120056141, b:120637874
TEST=emerge-octopus chromeos-bsp-baseboard-octopus
     chromeos-config-bsp-octopus chromeos-config chromeos-bsp-octopus

Change-Id: I0d786a18a0a12beff34ec02cb508cbd4d8749a0c
Reviewed-on: https://chromium-review.googlesource.com/1372169
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>