octopus: Add card name to UCM configs
This change explicitly adds the card name to the directory name for UCM
configs. The UCM card name is dictated by the kernel (except for the
suffix we add), and it receives a "sof-" prefix when SOF is enabled.

This change allows us to add "sof-" versions of these configurations in
parallel so that the SOF configurations can live in this package as
well.

BUG=b:120056141, b:120637874
TEST=emerge-octopus chromeos-bsp-baseboard-octopus
     chromeos-config-bsp-octopus chromeos-config chromeos-bsp-octopus

Change-Id: I636d5c1ac3df0b99e45ec11a518dd9b330cb8c34
Reviewed-on: https://chromium-review.googlesource.com/1372170
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>