octopus: Move SOF UCM settings out of baseboard
These UCM files were always installed by the baseboard package, and were
being used by systems where SOF was enabled. This unintentionally broke
audio for Meep (b/120637874) and caused the audio yaml file to not
reflect reality.

This change moves the SOF- based settings to the common overlay, and
updates the $card entry so that the proper settings will be used for
platforms that use glkda7219.

This leaves Meep still broken (as it has a ucm suffix that gets added),
but that will be fixed up a subsequent change which adds the proper UCM
settings for SOF for Meep.

Systems using the rt5682 codec are still broken after this change, but
we don't have proper UCM files for that configuration yet, so the $card
is not updated for those systems.

BUG=b:120056141, b:120637874
TEST=emerge-octopus chromeos-bsp-baseboard-octopus
     chromeos-config-bsp-octopus chromeos-config chromeos-bsp-octopus

Change-Id: Id95cbf9b6a5140d0ffab2c935bc7ee947826684c
Reviewed-on: https://chromium-review.googlesource.com/1372171
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>