kalista: add udev rules to whitelist the usb touchscreen as internal device
Currently, the usb device will be categorized as the external input device.
However, karma (a chromebase project) uses the USB touchsreen controller to
support large screen monitor (for example 23.8"). As a result, we need to
add the udev rules to whitelist the usb touchscreen as the internal device.
This patch also adds a rule to enable the touchscreen wakeup event.

Note: The ui doesn't support the udev now. This patch is added in case the
ui changes its implementation in the future.

BUG=b:120674137
TEST=emerge-kalista chromeos-bsp-kalista
     the udev rules were installed correctly.

Change-Id: Ic04b5e7d7794e7a0d34f727f03aca65f771daa23
Reviewed-on: https://chromium-review.googlesource.com/1390065
Commit-Ready: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>