[moblab] Add docker to moblab
To run a grpc-web to grpc backend you have to go through
a proxy, this deals with CORS and message translation.

From : https://grpc.io/blog/grpc-web-ga

"From the get-go, gRPC-Web will support Envoy"

To install envoy you have to use a docker image, I have
tried building from scratch but the bazel build files are
not well maintained and I never could get it to compile.

This is a almost standard install from gentoo, one patch
was added rootdir.patch as by default the directory chosen
if for some config tmp files is in a read only partition,
there is no way to pass a param or set an enviroment variable
to change this path.  So the code was changed from /etc/docker
to /var/run/docker

On the roadmap for grpc-web is an implemention where a proxy
will not be required, so this is a mid term solution and will
be removed when the proxy it no longer required.

BUG=chromium:682415
TEST=local moblab testing

Change-Id: I9ce39493ec6d7d4c6222d66ad41078449fb0db7d
Reviewed-on: https://chromium-review.googlesource.com/1504677
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>