lakitu: Disable /dev/mem
/dev/mem gives access to the machine's physical memory.
This level of access appears to be unnecessary for us, so
let's disable it. The upstream kernel recommends that
this config be disabled.

I believe this was originally added to ChromeOS for
firmware updates, but I'm really not sure and the
evidence is weak. At any rate, I don't think we
need this.

TEST=tryjob
BUG=b:116321333
RELEASE_NOTE=Disabled the CONFIG_DEVMEM configuration option
in the kernel.
CQ-DEPEND=CL:*780954

Change-Id: I168cc08bc2d322ccb5bad991ef0e29c7fac70e5a
Reviewed-on: https://chromium-review.googlesource.com/1404355
Commit-Ready: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Xuewei Zhang <xueweiz@google.com>