lakitu: sosreport: Add support for kdump.
This includes three changes:

1. Make "kdump" plugin collects the latest kernel dump
when config "kdump.latest_dump" is on, and collects all
kernel dumps when config "kdump.all_dumps" is on.

2. Make "release" plugin collect /etc/lsb-release and
/etc/os-release. This is used to find the COS version,
so that we can find the vmlinux image for kernel dump
debugging.

3. Set "kdump.latest_dump" to on in the default sosreport
configuration on COS.

This is submitted to upstream at:
https://github.com/sosreport/sos/pull/1549

BUG=b:115771082
TEST=validation test
RELEASE_NOTE=None

Change-Id: Ib3143b66efc1ee3a737c31643800f75c3dbcd7c0
Reviewed-on: https://chromium-review.googlesource.com/1444496
Commit-Ready: Xuewei Zhang <xueweiz@google.com>
Tested-by: Xuewei Zhang <xueweiz@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>