chipset-skl: fix performance ranges which exceed 2.2x tolerance
The performance range rate (upper-limit / lower-limit) is restricted to be within the
tolerance(=2.2), tests may fail if the published range rate is violated.

In this CL, we regard Lars as the lowest performance device of chipset-skl, and modify
some lower-limit values within the tolerance rate and at the same time could cover the
measured values from Lars.

For ARC.* performance we have found some of the measured values of Lars are even lower.
For those cases we re-strike the upper-limit as
sqrt(prev-upper-limit * lars-lower-limit * 2.2) to cover both of them.

BUG=b:123731960
TEST=build Sentry and Lars image

Change-Id: Ia0a39c98cac4028dcb646d64403d847d331f0724
Reviewed-on: https://chromium-review.googlesource.com/1466464
Commit-Ready: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>