mt8173: remove LDFLAG --fix-cortex-a53-843419
This flag was added to enable an errata for the little core (A53) on
MT8173 chips.  However, this errata only affects arm64 linking, not
arm32.  The entire userspace for the Chromium OS mt8173 build is actually
arm32, only the kernel is built for arm64.  Since the kernel handles
this flag itself, via the CONFIG_ARM64_ERRATUM_843419 KConfig option,
this LDFLAG need not be set in the chipset overlay.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:929415,b:35550539
TEST=hana / elm trybot

Change-Id: Id893edd0c9119ddc9d0272b9db8612f074ee6a9d
Reviewed-on: https://chromium-review.googlesource.com/1473490
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>