Hatch: Enable cros-pd, eclog, VMs
Enable some of the options that we were missing.

cros-pd enables the chromium power delivery instead of the ACPI power.
eclog enables us to get ec logs in the OS.
The rent enables VMs. Not sure if this is useful, but since it is part
of nami build, it must be.

BUG=b:123541187
TEST=Test that the /sy/class/power_supply directory shows the cros pd
chargers.
Change-Id: I4dbfa3efd53473605294da166ddd513ea017e3b8
Reviewed-on: https://chromium-review.googlesource.com/1474613
Commit-Ready: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>
Reviewed-by: Benson Leung <bleung@chromium.org>