Revert "lakitu: Add app-admin/sosreport to base image."
This reverts commit bb574b8a10534b4166f9b51b4e6441600cec8a25.

Reason for revert:
Sorry folks, turns out the latest 69 builder failure is caused by installing app-admin/sosreport (i.e. this patch):
https://cros-goldeneye.corp.google.com/chromeos/healthmonitoring/buildDetails?buildbucketId=8920488779018356192

The failure reason is that sosreport installs to /usr/share/locale, which is blacklisted in CrOS.

I verified that "emerge-lakitu sosreport" would create that blacklisted directory in 69, but won't create it in 70/HEAD.

I do not yet understand why is that the case. But let's revert this commit to unblock our 69 daily build.

Original change's description:
> lakitu: Add app-admin/sosreport to base image.
> 
> BUG=b:80099316
> TEST=Manual
> RELEASE_NOTE=None
> 
> CQ-DEPEND=CL:1480571
> 
> Change-Id: If919409b1f4130b02cda970fe282735a86995da4
> Reviewed-on: https://chromium-review.googlesource.com/1200248
> Commit-Ready: Xuewei Zhang <xueweiz@google.com>
> Tested-by: Xuewei Zhang <xueweiz@google.com>
> Reviewed-by: Pradeep Sawlani <sawlani@google.com>
> (cherry picked from commit 906567a2620016a047b89b794467ecffc2224327)

Bug: b:80099316
Change-Id: I5f5b71009097c5d207d24d7f0cc72f4a4823cca1