lakitu: include 'perf' in the base image
'perf' is critical to debug certain prod issues and since it is tightly
coupled with the kernel version, it is difficult to install & run it
from a docker container. So include in the base image.

BUG=b:30671224
TEST=manually verified perf can be used on the base image
RELEASE_NOTE=Included 'perf' tool in the image.

Change-Id: I254542f004d48b24e29bc9f4e271c5652676351d
(cherry picked from commit 9722bf52027fc4509c5978594ce99093d124fe75)