falco: increase rootfs size to 1400MB
falco_li is running out of space on the rootfs. It only has 11MB out of
1240MB free space left. However, the partition is 2GB large and falco_li
will be EOL in < 6 months, so future growth in rootfs will be limited.
We can simply bump the rootfs size to 1400MB, it should be enough for
the supported lifetime of the device.

BUG=chromium:943908
TEST=Pre-CQ passes, rootfs size is not 1400MiB (358400 blocks of 4096
bytes)

Change-Id: I048c094dcd4a3fc45df84eb55db7c85e4e267c8c
Reviewed-on: https://chromium-review.googlesource.com/1539294
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>