scripts: disable loadpin module in factory
Touchscreen driver cannot load a firmware file that is not in rootfs.
To make it easier to maintain factory binary, disable "LoadPin" so
factory can easily modify test image and install shim to use a
touchscreen firmware outside of factory.  (This is especially useful for
factory install shim, because the install shim needs to be signed, in
this casae, disable rootfs verification and modify rootfs doesn't work).

BUG=b:129240949
TEST=local build images and check kernel command line

Change-Id: I50b74bb65991860356172268170af6b6de350aae