Enable Kiosk Next for scarlet boards.
This new feature will only be available in scarlet boards for now, so the plan is to
add this USE flag to scarlet and add a new gn arg when the USE flag is set when
compiling Chrome (crrev/r/1566310).

A gn arg will be added based on the existence of this USE flag, that work is being
done in crrev/r/1566310.

BUG=chromium:933174
TEST=Built a scarlet image and verified that the enable_kiosk_next arg is being passed.

Change-Id: I7acc219def4db071092e6e2c16759fbbf6386277
Reviewed-on: https://chromium-review.googlesource.com/1566062
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>