Fully specify UCM config for samus-kernelnext
Previously, samus relied on matching to determine device configuration
in cras. However, a kernel change between 3.14 and 4.14 means that the
automatic matching results in the incorrect jack being associated with
the Headphone output, which breaks use case switching.

Convert samus-kernelnext to use a fully specified UCM config so that
the proper jack is selected and that headphone/audio switching works
properly. After samus kernel uprev is complete, the split audio
configs can be removed.

BUG=b:130221745, b:130557299

TEST=run cras with updated UCM config, check that input and output
switching works properly, volume is appropriately loud, and that the
proper mixer controls are enabled when headphones are inserted.

Change-Id: I86b482e9f2bfc086ed027d3207b2825d247a0487
Reviewed-on: https://chromium-review.googlesource.com/1588677
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: En-Shuo Hsu <enshuo@google.com>