Disable USE=dt_compression on old Arm boards
This patch introduces the new dt_compression USE flag that will be used
by the kernel ebuilds to decide whether to compress flattened device
trees in the FIT image. This feature is generally useful for all device
tree boards and so will be enabled by default for all boards (to avoid
future board overlays forgetting to enable it individually). However,
firmware on already shipped Arm boards does not support this feature,
and it is not really feasible to backport. Therefore, disable the new
default USE flag on those boards.

BUG=None
TEST=Built Kevin, confirmed FDTs were still uncompressed.

Change-Id: I1a40a3ea8f73aa16d965716e3dec51fd60ae518f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1621519
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>