sarien: Add upstart script to continually sync EC's RTC
The EC on Wilco contains an RTC that needs to be kept in sync
with the local time (not UTC, like most linux RTCs) so that it
can properly control the charge schedule. This upstart script
will re-sync the EC's RTC to the local time every 5 minutes.
Since the charge scheduling only uses 15 minute granularity,
this should be frequent enough.

BUG=b:132692517, b:127678805, chromium:939132
TEST=manual: Set the RTC to some bogus time with
$ hwclock --set --date="01 JAN 2001 01:01:01" --rtc=/dev/rtc1
and 5 minutes later it is corrected per
$ hwclock --show --rtc=/dev/rtc1
The time is also adjusted at
$ start wilco_sync_ec_rtc
and
$ stop wilco_sync_ec_rtc
and the job restarts itself if you kill the process.

Cq-Depend: chromium:1653915
Change-Id: I5e603066585f9c6ccba24e0bb82cd2e0eb42233b
Signed-off-by: Nick Crews <ncrews@chromium.org>
(cherry picked from commit 240a04c2292e4f4b03951135737117775aabc3ba)