lakitu: Remove checking of uid and gid to be same
Recently, when update engine tries to update an instance
from old COS image to newer COS image, update gets failed.
This happens due to following reason:

1. When google-account-daemon tries to allocate UID and GID
   for users that have existing /home/<user> dir, it tries to use
   old UID and GID if old UID and GID were same.
2. Recently, a new user 'crash' had been added with UID and GID as
   20137.

So, after the update, when google-accounts-daemon tries to allocate
UID and GID for existing users, it is not able to use the old UID
and GID as they are not same which is a mandate for the associated
patch.Thereby, it assigns a new UID and GID which is greater than
20137 (above exisiting highest UID). This results in bad ownership
while doing ssh as the UID and GID associated with /home/<user> is
different from UID and GID of user.

BUG=b:135045945
TEST=deleting user from /etc/passwd and /etc/group and trying ssh
RELEASE_NOTE=None

Change-Id: I01b69fd4cb9c51cd290ab17dd31a5e02ed4a65d2
Reviewed-on: https://chromium-review.googlesource.com/1668947
Tested-by: Vaibhav Rustagi <vaibhavrustagi@google.com>
Commit-Ready: Vaibhav Rustagi <vaibhavrustagi@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Daniel Wang <wonderfly@google.com>