kukui: add decoder media device udev rule
The use of the request API with the decoder requires us to open the
decoder media device in order to create and submit requests. However,
media devices have the same shortcomings as video devices, in that they
are just sequentially named media0, media1, ... without any indication
about which is which.

This CL identifies the decoder media devices and creates a media-dec0
symlink to it, similarly to the way we already create a video-dec0
symlink to the video decoder device. That way Chromium can open the
correct media device without any hassle.

BUG=b:123551776
TEST=Checked that /dev/media-dec0 was visible upon boot.

Change-Id: If76990159ca3ecf11ef9bb50256702fb9662230c
Reviewed-on: https://chromium-review.googlesource.com/1675445
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Ready: Alexandre Courbot <acourbot@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>