veyron: Disable clang / afdo for all veyron for 4.19
In <https://crrev.com/i/1221451> we had disabled this just for minnie,
but really this rule should be the same for all veyron devices trying
to compile 4.19.  Then if/when we are able to use clang we can just
switch it in one place.

BUG=chromium:819808, chromium:941638
TEST=Building 4.19 gives the right compiler

Change-Id: I9553f641ab8366d946d87b2b8cb7bfa135c31d43
Reviewed-on: https://chromium-review.googlesource.com/1680445
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>