lakitu: system-sysdaemons: remove NPD service from the globally controlled list
system-sysdaemons provides an unified way of adding daemons to
system-sysdaemons.slice. However, it is not necessary, as adding below
line in the service file has equal effect:
Slice=system-sysdaemons.slice

In order to make node-problem-detector.service to not show up in Lakitu,
we can temporarily remove it from system-sysdaemons package.

(Another solution would be to implement an IUSE flag in this package,
which seems like an overkill.)

BUG=b:134962379,b:32505363
TEST=None
RELEASE_NOTE=None

Change-Id: I03aafbc98e3a65d8d6a4e125cf74303760a087b1