Add chipset-qc7180 overlay, rebase baseboard-trogdor to it
This is the overlay for the Qualcomm SC7180. Name it qc7180 instead
since for some reason that seems to be how we prefix all Qualcomm
overlays.

Just copying everything from qc845 for now, can be fixed later if
necessary. Dropped the testclock utility -- if we decide that we also
need that here, we should probably find a better home for it.

NOTE: Everyone who has already run setup_board --board=trogdor will have
to delete /build/trogdor and rerun it after syncing this patch!

BUG=b:135696461
TEST=None

Change-Id: I745a5423860921a84a347591a604f2bdd2dc9d39
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1700172
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>