auron_paine: Swapped mic input sources
Between 3.14 and 4.14 the UCM mappings for the input sources swapped
between the built-in microphone and the external jack. This has the
effect of enabling the external jack by default and enabling the
built-in microphone when an external mic is connected.

Rather than tracking down the exact reason for this, simply swap the
devices in the UCM configs since the configs reside in the overlays now.

BUG=b:133861656
TEST=Check both speakers and microphone in Chrome and verify the sources
are correct for both the built-in mic and external jack.

Change-Id: Ie394817f1d38b129299c2eeb92d270afa979c7a3
Reviewed-on: https://chromium-review.googlesource.com/1702854
Tested-by: Jon Flatley <jflat@chromium.org>
Commit-Ready: Jon Flatley <jflat@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>